cooler_jazzmin/helpers/server_util/requirements.txt
GPUtil==1.4.0
psutil==5.9.4
psycopg2-binary
PyYAML==6.0
